Update examples for output devices
Rename some old examples that are for capture devices and make new versions of those examples for output devices.
